requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.18.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

